+++ CHANGELOG.md
@@ -7,8 +7,12 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## [Unreleased][unreleased]
 -
 
-## 1.0.0 - 2016-06-16
+## [1.0.1] - 2016-10-20
+- Added transaction feedback status setting.
+- Fixed - Too many arguments for function "__".
+
+## 1.0.0 - 2016-07-06
 - First release.
 
-[unreleased]: https://github.com/wp-pay-gateways/ems-e-commerce/compare/1.0.0...HEAD
+[unreleased]: https://github.com/wp-pay-gateways/ems-e-commerce/compare/1.0.1...HEAD
 [1.0.1]: https://github.com/wp-pay-gateways/ems-e-commerce/compare/1.0.0...1.0.1